etl/requirements.txt
pandas==2.2.2
python-dateutil==2.9.0
openpyxl==3.1.5